Separação dos scheduled
